requirements.txt
discord.py==1.5.1
SQLAlchemy==1.3.22
PyMySQL==0.10.1